+++ common/types.go
@@ -85,6 +85,12 @@ var (
 
 	// ErrEmptyJSONHTTPResponse is returned when the JSONHTTPResponse is nil.
 	ErrEmptyJSONHTTPResponse = errors.New("empty json http response")
+
+	// ErrEmptyResponse is returned when the response body is empty.
+	ErrEmptyResponse = errors.New("empty response body")
+
+	// ErrRecordDataNotJSON is returned when the record data in WriteParams is not JSON.
+	ErrRecordDataNotJSON = errors.New("record data is not JSON")
 )
 
 // ReadParams defines how we are reading data from a SaaS API.
