--- outreach/model.go
@@ -1,9 +1,0 @@
-package outreach
-
-const (
-	attributesKey    string = "attributes"
-	relationshipsKey string = "relationships"
-	idKey            string = "id"
-	typeKey          string = "type"
-	dataKey          string = "data"
-)
