--- outreach/provider.go
@@ -1,8 +1,0 @@
-package outreach
-
-import "github.com/amp-labs/connectors/providers"
-
-// Provider returns the connector provider.
-func (c *Connector) Provider() providers.Provider {
-	return providers.Outreach
-}
