--- outreach/string.go
@@ -1,5 +1,0 @@
-package outreach
-
-func (c *Connector) String() string {
-	return c.Provider() + ".Connector"
-}
